# storage: reuse byte buffer for `PointSynthesizingIter.UnsafeRawMVCCKey`